Move validation to compiler package.
