+++ package.json
@@ -21,6 +21,7 @@
     "@milesj/build-tools": "^1.0.1",
     "@types/extend": "^3.0.1",
     "@types/hoist-non-react-statics": "^3.3.1",
+    "@types/js-yaml": "^3.11.1",
     "@types/prop-types": "^15.7.3",
     "@types/react": "^16.9.19",
     "@types/react-dom": "^16.9.5",
