+++ packages/compiler/.eslintrc
@@ -1,0 +1,6 @@
+{
+  "rules": {
+    "no-magic-numbers": "off",
+    "sort-keys": "off"
+  }
+}
