+++ packages/compiler/.npmignore
@@ -1,0 +1,7 @@
+src/
+tests/
+types/
+*.log
+*.tsbuildinfo
+.eslintrc
+tsconfig.json
