+++ packages/compiler/README.md
@@ -1,0 +1,13 @@
+# Design System Framework
+
+[![Build Status](https://github.com/milesj/aesthetic/workflows/Build/badge.svg)](https://github.com/milesj/aesthetic/actions?query=branch%3Amaster)
+[![npm version](https://badge.fury.io/js/%40aesthetic%framework.svg)](https://www.npmjs.com/package/@aesthetic/framework)
+[![npm deps](https://david-dm.org/milesj/aesthetic.svg?path=packages/framework)](https://www.npmjs.com/package/@aesthetic/framework)
+
+A framework for building a design system.
+
+## Installation
+
+```
+yarn add @aesthetic/framework
+```
