+++ packages/compiler/package.json
@@ -1,14 +1,14 @@
 {
-  "name": "@aesthetic/framework",
+  "name": "@aesthetic/compiler",
   "version": "0.0.0",
-  "description": "A design system framework.",
+  "description": "Compile a design system configuration to a specified platform target.",
   "keywords": [
     "aesthetic",
     "design",
     "system",
-    "language"
+    "compiler"
   ],
-  "repository": "https://github.com/milesj/aesthetic/tree/master/packages/framework",
+  "repository": "https://github.com/milesj/aesthetic/tree/master/packages/compiler",
   "license": "MIT",
   "main": "./lib/index.js",
   "module": "./esm/index.js",
@@ -22,8 +22,8 @@
     "url": "https://ko-fi.com/milesjohnson"
   },
   "dependencies": {
-    "@aesthetic/sss": "^0.0.0",
-    "aesthetic-utils": "^3.0.0",
-    "extend": "^3.0.2"
+    "@aesthetic/framework": "^0.0.0",
+    "js-yaml": "^3.13.1",
+    "optimal": "^4.0.0"
   }
 }
