+++ packages/compiler/src/index.ts
@@ -1,0 +1,5 @@
+import ConfigLoader from './ConfigLoader';
+
+export * from './types';
+
+export { ConfigLoader };
