+++ packages/compiler/src/types.ts
@@ -1,0 +1,5 @@
+import { DesignConfig, ThemeConfig } from '@aesthetic/framework';
+
+export interface ConfigFile extends DesignConfig<string> {
+  themes: { [name: string]: ThemeConfig<string> };
+}
